+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## 1.0.10
 - When migrating files in place, don't replace the original file if no conversion has taken place. Based on PR[#78] by Semiao Marco.
+- When converting a file in an archive, update the last modified time for that archive entry.  Based on PR[#78] by Semiao Marco.
 
 ## 1.0.9
 - Update the JaCoCo Maven plugin to 0.8.12. (dependabot/markt)